+++ ui/analyse/src/view/clocks.ts
@@ -3,6 +3,7 @@ import type AnalyseCtrl from '../ctrl';
 import { defined, notNull } from 'lib';
 import * as licon from 'lib/licon';
 import { iconTag } from 'lib/snabbdom';
+import { formatNvuiClockTime } from 'lib/game/clock/clockView';
 
 interface ClockOpts {
   centis: number | undefined;
@@ -59,7 +60,11 @@ export default function renderClocks(ctrl: AnalyseCtrl, path: Tree.Path): [VNode
 }
 
 const renderClock = (opts: ClockOpts): VNode =>
-  h('div.analyse__clock.' + opts.cls, { class: { active: opts.active } }, site.blindMode ? clockContentNvui(opts) : clockContent(opts) );
+  h(
+    'div.analyse__clock.' + opts.cls,
+    { class: { active: opts.active } },
+    site.blindMode ? clockContentNvui(opts) : clockContent(opts),
+  );
 
 function clockContent(opts: ClockOpts): Array<string | VNode> {
   if (!opts.centis && opts.centis !== 0) return ['-'];
@@ -79,23 +84,7 @@ function clockContent(opts: ClockOpts): Array<string | VNode> {
 
 function clockContentNvui(opts: ClockOpts): Array<string | VNode> {
   if (!opts.centis && opts.centis !== 0) return ['None'];
-  const totalSeconds = Math.floor(opts.centis / 100);
-  const hours = Math.floor(totalSeconds / 3600);
-  const minutes = Math.floor((totalSeconds % 3600) / 60);
-  const seconds = totalSeconds % 60;
-
-  const parts: string[] = [];
-  if (hours > 0) parts.push(i18n.site.nbHours(hours));
-  if (minutes > 0) parts.push(i18n.site.nbMinutes(minutes));
-  if (seconds > 0 || parts.length === 0) parts.push(i18n.site.nbSeconds(seconds));
-
-
-  const pauseStr = opts.pause ? ' - Paused' : '';
-  if (parts.length === 1) {
-    return [parts[0] + pauseStr];
-  } else {
-    return [parts.join(', ') + pauseStr];
-  }
+  return [formatNvuiClockTime(opts.centis * 10)];
 }
 
 const pad2 = (num: number): string => (num < 10 ? '0' : '') + num;
